requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
plotly==5.22.0
statsmodels==0.14.1
scikit-learn==1.4.2
networkx==3.3
PyYAML==6.0.2
tqdm==4.66.4
# streamlit is optional but included for the dashboard
streamlit==1.36.0
